+++ src/main/java/com/atlassian/db/replica/api/DualConnection.java
@@ -8,7 +8,7 @@ import com.atlassian.db.replica.internal.circuitbreaker.BreakerConnection;
 import com.atlassian.db.replica.spi.ConnectionProvider;
 import com.atlassian.db.replica.spi.DatabaseCall;
 import com.atlassian.db.replica.spi.ReplicaConsistency;
-import com.atlassian.db.replica.spi.circuitbreaker.CircuitBreaker;
+import com.atlassian.db.replica.spi.CircuitBreaker;
 import com.atlassian.db.replica.spi.state.StateListener;
 
 import java.sql.*;