+++ src/main/java/com/atlassian/db/replica/api/SqlRun.java
@@ -1,0 +1,7 @@
+package com.atlassian.db.replica.api;
+
+import java.sql.SQLException;
+
+public interface SqlRun {
+    void run() throws SQLException;
+}