+++ src/main/java/com/atlassian/db/replica/internal/circuitbreaker/BreakerConnection.java
@@ -1,7 +1,8 @@
 package com.atlassian.db.replica.internal.circuitbreaker;
 
 import com.atlassian.db.replica.api.SqlCall;
-import com.atlassian.db.replica.spi.circuitbreaker.CircuitBreaker;
+import com.atlassian.db.replica.spi.CircuitBreaker;
+import com.atlassian.db.replica.api.SqlRun;
 
 import java.sql.Array;
 import java.sql.Blob;
@@ -73,7 +74,7 @@ public class BreakerConnection implements Connection {
 
     @Override
     public void rollback() throws SQLException {
-        breaker.handle((CircuitBreaker.SqlRunnable) delegate::rollback);
+        breaker.handle((SqlRun) delegate::rollback);
     }
 
     @Override