+++ src/main/java/com/atlassian/db/replica/internal/circuitbreaker/BreakerStatement.java
@@ -1,7 +1,7 @@
 package com.atlassian.db.replica.internal.circuitbreaker;
 
 import com.atlassian.db.replica.api.SqlCall;
-import com.atlassian.db.replica.spi.circuitbreaker.CircuitBreaker;
+import com.atlassian.db.replica.spi.CircuitBreaker;
 
 import java.sql.Connection;
 import java.sql.ResultSet;