+++ src/main/java/com/atlassian/db/replica/internal/circuitbreaker/ClosedBreaker.java
@@ -1,7 +1,8 @@
 package com.atlassian.db.replica.internal.circuitbreaker;
 
 import com.atlassian.db.replica.api.SqlCall;
-import com.atlassian.db.replica.spi.circuitbreaker.CircuitBreaker;
+import com.atlassian.db.replica.spi.CircuitBreaker;
+import com.atlassian.db.replica.api.SqlRun;
 
 import java.sql.SQLException;
 
@@ -18,7 +19,7 @@ public class ClosedBreaker implements CircuitBreaker {
     }
 
     @Override
-    public void handle(SqlRunnable runnable) throws SQLException {
-        runnable.run();
+    public void handle(SqlRun run) throws SQLException {
+        run.run();
     }
 }