+++ CHANGELOG.md
@@ -28,9 +28,13 @@ Dropping a requirement of a major version of a dependency is a new contract.
 
   It only asks if it can be called and how to handle calls.
   This lets you track any state and wire in advanced CircuitBreaker libs.
+- Move `CircuitBreaker` from `spi.circuitbreaker` to `spi`.
 
 ### Removed
 - Remove `BreakerState`.
+
+### Added
+- Add `SqlRun` required by `CircuitBreaker`.
 
 ## [0.1.27] - 2021-02-23
 [0.1.27]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.26...release-0.1.27